# Histogram Aggregation: Fix bucket detection logic, performance improvements, and benchmark tests