--- a/sdk/include/opentelemetry/sdk/metrics/aggregation/aggregation_config.h
+++ b/sdk/include/opentelemetry/sdk/metrics/aggregation/aggregation_config.h
@@ -3,8 +3,10 @@
 
 #pragma once
 
-#include <list>
 #include "opentelemetry/version.h"
+
+#include <vector>
+
 OPENTELEMETRY_BEGIN_NAMESPACE
 namespace sdk
 {
@@ -19,7 +21,7 @@ public:
 class HistogramAggregationConfig : public AggregationConfig
 {
 public:
-  std::list<double> boundaries_;
+  std::vector<double> boundaries_;
   bool record_min_max_ = true;
 };
 }  // namespace metrics