--- a/sdk/include/opentelemetry/sdk/metrics/data/point_data.h
+++ b/sdk/include/opentelemetry/sdk/metrics/data/point_data.h
@@ -8,7 +8,7 @@
 #include "opentelemetry/sdk/metrics/instruments.h"
 #include "opentelemetry/version.h"
 
-#include <list>
+#include <vector>
 
 OPENTELEMETRY_BEGIN_NAMESPACE
 namespace sdk
@@ -55,14 +55,14 @@ public:
   HistogramPointData &operator=(HistogramPointData &&) = default;
   HistogramPointData(const HistogramPointData &)       = default;
   HistogramPointData()                                 = default;
-  HistogramPointData(std::list<double> &boundaries) : boundaries_(boundaries) {}
-  std::list<double> boundaries_ = {};
-  ValueType sum_                = {};
-  ValueType min_                = {};
-  ValueType max_                = {};
-  std::vector<uint64_t> counts_ = {};
-  uint64_t count_               = {};
-  bool record_min_max_          = true;
+  HistogramPointData(std::vector<double> &boundaries) : boundaries_(boundaries) {}
+  std::vector<double> boundaries_ = {};
+  ValueType sum_                  = {};
+  ValueType min_                  = {};
+  ValueType max_                  = {};
+  std::vector<uint64_t> counts_   = {};
+  uint64_t count_                 = {};
+  bool record_min_max_            = true;
 };
 
 class DropPointData