--- a/sdk/test/metrics/BUILD
+++ b/sdk/test/metrics/BUILD
@@ -48,6 +48,22 @@ cc_test(
 )
 
 cc_test(
+    name = "histogram_test",
+    srcs = [
+        "histogram_test.cc",
+    ],
+    tags = [
+        "metrics",
+        "test",
+    ],
+    deps = [
+        "//sdk/src/metrics",
+        "//sdk/src/resource",
+        "@com_google_googletest//:gtest_main",
+    ],
+)
+
+cc_test(
     name = "view_registry_test",
     srcs = [
         "view_registry_test.cc",