--- a/sdk/test/metrics/aggregation_test.cc
+++ b/sdk/test/metrics/aggregation_test.cc
@@ -76,17 +76,17 @@ TEST(Aggregation, LongHistogramAggregation)
   ASSERT_TRUE(nostd::holds_alternative<int64_t>(histogram_data.sum_));
   EXPECT_EQ(nostd::get<int64_t>(histogram_data.sum_), 0);
   EXPECT_EQ(histogram_data.count_, 0);
-  aggr.Aggregate((int64_t)12, {});   // lies in fourth bucket
-  aggr.Aggregate((int64_t)100, {});  // lies in eight bucket
+  aggr.Aggregate((int64_t)12, {});   // lies in third bucket
+  aggr.Aggregate((int64_t)100, {});  // lies in sixth bucket
   histogram_data = nostd::get<HistogramPointData>(aggr.ToPoint());
   EXPECT_EQ(nostd::get<int64_t>(histogram_data.min_), 12);
   EXPECT_EQ(nostd::get<int64_t>(histogram_data.max_), 100);
   EXPECT_EQ(nostd::get<int64_t>(histogram_data.sum_), 112);
   EXPECT_EQ(histogram_data.count_, 2);
   EXPECT_EQ(histogram_data.counts_[3], 1);
-  EXPECT_EQ(histogram_data.counts_[7], 1);
-  aggr.Aggregate((int64_t)13, {});   // lies in fourth bucket
-  aggr.Aggregate((int64_t)252, {});  // lies in ninth bucket
+  EXPECT_EQ(histogram_data.counts_[6], 1);
+  aggr.Aggregate((int64_t)13, {});   // lies in third bucket
+  aggr.Aggregate((int64_t)252, {});  // lies in eigth bucket
   histogram_data = nostd::get<HistogramPointData>(aggr.ToPoint());
   EXPECT_EQ(histogram_data.count_, 4);
   EXPECT_EQ(histogram_data.counts_[3], 2);
@@ -165,17 +165,17 @@ TEST(Aggregation, DoubleHistogramAggregation)
   ASSERT_TRUE(nostd::holds_alternative<double>(histogram_data.sum_));
   EXPECT_EQ(nostd::get<double>(histogram_data.sum_), 0);
   EXPECT_EQ(histogram_data.count_, 0);
-  aggr.Aggregate(12.0, {});   // lies in fourth bucket
-  aggr.Aggregate(100.0, {});  // lies in eight bucket
+  aggr.Aggregate(12.0, {});   // lies in third bucket
+  aggr.Aggregate(100.0, {});  // lies in sixth bucket
   histogram_data = nostd::get<HistogramPointData>(aggr.ToPoint());
   EXPECT_EQ(nostd::get<double>(histogram_data.sum_), 112);
   EXPECT_EQ(histogram_data.count_, 2);
   EXPECT_EQ(histogram_data.counts_[3], 1);
-  EXPECT_EQ(histogram_data.counts_[7], 1);
+  EXPECT_EQ(histogram_data.counts_[6], 1);
   EXPECT_EQ(nostd::get<double>(histogram_data.min_), 12);
   EXPECT_EQ(nostd::get<double>(histogram_data.max_), 100);
-  aggr.Aggregate(13.0, {});   // lies in fourth bucket
-  aggr.Aggregate(252.0, {});  // lies in ninth bucket
+  aggr.Aggregate(13.0, {});   // lies in third bucket
+  aggr.Aggregate(252.0, {});  // lies in eight bucket
   histogram_data = nostd::get<HistogramPointData>(aggr.ToPoint());
   EXPECT_EQ(histogram_data.count_, 4);
   EXPECT_EQ(histogram_data.counts_[3], 2);