--- a/sdk/test/metrics/aggregation_test.cc
+++ b/sdk/test/metrics/aggregation_test.cc
@@ -132,9 +132,9 @@ TEST(Aggregation, LongHistogramAggregationBoundaries)
 {
   std::shared_ptr<opentelemetry::sdk::metrics::HistogramAggregationConfig> aggregation_config{
       new opentelemetry::sdk::metrics::HistogramAggregationConfig};
-  std::list<double> user_boundaries = {0.0,   50.0,   100.0,  250.0,  500.0,
-                                       750.0, 1000.0, 2500.0, 5000.0, 10000.0};
-  aggregation_config->boundaries_   = user_boundaries;
+  std::vector<double> user_boundaries = {0.0,   50.0,   100.0,  250.0,  500.0,
+                                         750.0, 1000.0, 2500.0, 5000.0, 10000.0};
+  aggregation_config->boundaries_     = user_boundaries;
   LongHistogramAggregation aggr{aggregation_config.get()};
   auto data = aggr.ToPoint();
   ASSERT_TRUE(nostd::holds_alternative<HistogramPointData>(data));
@@ -146,9 +146,9 @@ TEST(Aggregation, DoubleHistogramAggregationBoundaries)
 {
   std::shared_ptr<opentelemetry::sdk::metrics::HistogramAggregationConfig> aggregation_config{
       new opentelemetry::sdk::metrics::HistogramAggregationConfig};
-  std::list<double> user_boundaries = {0.0,   50.0,   100.0,  250.0,  500.0,
-                                       750.0, 1000.0, 2500.0, 5000.0, 10000.0};
-  aggregation_config->boundaries_   = user_boundaries;
+  std::vector<double> user_boundaries = {0.0,   50.0,   100.0,  250.0,  500.0,
+                                         750.0, 1000.0, 2500.0, 5000.0, 10000.0};
+  aggregation_config->boundaries_     = user_boundaries;
   DoubleHistogramAggregation aggr{aggregation_config.get()};
   auto data = aggr.ToPoint();
   ASSERT_TRUE(nostd::holds_alternative<HistogramPointData>(data));