Fixes #1864, #1865, #1866

## Changes

1. Fixed the bucket detection logic
    -  Lower bound value is exclusive, Upper bound value is inclusive ( as mentioned in [specs ](https://github.com/open-telemetry/opentelemetry-specification/blob/main/specification/metrics/sdk.md#explicit-bucket-histogram-aggregation))
    - The last bucket was not getting populated.
3. The default boundaries for UINT64 was incorrect, changed it to use the values as specified in above specs link
4. Added separate tests for histogram buckets with default and custom boundaries (shamelessly taken from the issue).
5. Pass custom boundaries config data while merging and diff-ing Histogram aggregation.
6. Change boundaries data type from std::list to std::vector. And use binary search for searching the right bucket for a measurement.
7. Add histogram performance benchmark tests.

Benchmark result for histogram aggregation (prior to changes in point 5 above):
```
$ ./histogram_aggregation_benchmark
2022-12-16 15:18:01
Running ./histogram_aggregation_benchmark
Run on (20 X 2808.01 MHz CPU s)
CPU Caches:
  L1 Data 32K (x10)
  L1 Instruction 32K (x10)
  L2 Unified 256K (x10)
  L3 Unified 20480K (x1)
Load Average: 2.90, 1.64, 0.61
------------------------------------------------------------------
Benchmark                        Time             CPU   Iterations
------------------------------------------------------------------
BM_HistogramAggregation        740 ns          740 ns       936807
```
Benchmark result for histogram aggregation (after changes in point 5 above):
```
$ ./histogram_aggregation_benchmark
2022-12-16 15:17:58
Running ./histogram_aggregation_benchmark
Run on (20 X 2808.01 MHz CPU s)
CPU Caches:
  L1 Data 32K (x10)
  L1 Instruction 32K (x10)
  L2 Unified 256K (x10)
  L3 Unified 20480K (x1)
Load Average: 2.90, 1.64, 0.61
------------------------------------------------------------------
Benchmark                        Time             CPU   Iterations
------------------------------------------------------------------
BM_HistogramAggregation        552 ns          552 ns      1297628
```





For significant contributions please make sure you have completed the following items:

* [ ] `CHANGELOG.md` updated for non-trivial changes
* [x] Unit tests have been added
* [x] Changes in public API reviewed