@ThomsonTan 2023-01-05
> Can I merge this PR, as this has few trivial bug fixes, and performance improvement for histogram aggregation, don't want to keep this on hold for long. Though it is approved, I realize it is still not meeting our merge policy, so checking if there are any issue :) I can incorporate any later comments as separate PR if required.
> 
> `A pull request opened by an Approver / Maintainer can be merged with only one approval from Approver / Maintainer (at different company).`
> 
> p.s. - As special case, we have been merging ETW exporter related PRs and few smaller PRs with one approval from same company, but that should be acceptable.

Thanks for raising this. @esigo could you please help take a look when you have time?

For ETW exporter, it seems more appropriate to move it to contrib repo.